requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
pydantic
email_validator
python-dotenv
python-jose
python-multipart